Fix migration adding portal type to news and feedback

+++ UJCSystem/Vokabular.DataEntities/Database/Entities/Enums/PortalTypeEnum.cs
@@ -1,6 +1,6 @@
 ﻿namespace Vokabular.DataEntities.Database.Entities.Enums
 {
-    public enum PortalTypeEnum : byte
+    public enum PortalTypeEnum : short
     {       
         Research = 0,
         Community = 1,
